Updates the public address to *url.URL
Make the "listen" and "public" addresses more explicit: they both must be resolved to an `url.URL`. "listen" must be of form "tcp://<host>:<post>" and "public" to "//<host>:<port>/<path>". Right now we don't take the scheme into account, but could be the case later with "ipv4://1.2.3.4:0/path" or "dns://example.com:443/path". However it would require to implement the resolvers for gRPC.

- updates the certificate creation to check if it's an IP or a DNSName
- adds tests to improve coverage
- updates the doc